+++ color/src/flags.rs
@@ -56,11 +56,7 @@ impl Flags {
 
     /// Set the missing components.
     #[inline]
-    #[warn(
-        clippy::missing_const_for_fn,
-        reason = "can be made const with MSRV 1.83"
-    )]
-    pub fn set_missing(&mut self, missing: Missing) {
+    pub const fn set_missing(&mut self, missing: Missing) {
         self.missing = missing;
     }
 
@@ -91,11 +87,7 @@ impl Flags {
 
     /// Set the flags to indicate the color was specified using one of the named color space
     /// functions.
-    #[warn(
-        clippy::missing_const_for_fn,
-        reason = "can be made const with MSRV 1.83"
-    )]
-    pub(crate) fn set_named_color_space(&mut self) {
+    pub(crate) const fn set_named_color_space(&mut self) {
         self.name = 255;
     }
 
@@ -120,11 +112,7 @@ impl Flags {
 
     /// Discard the color name or color space name from the flags.
     #[inline]
-    #[warn(
-        clippy::missing_const_for_fn,
-        reason = "can be made const with MSRV 1.83"
-    )]
-    pub fn discard_name(&mut self) {
+    pub const fn discard_name(&mut self) {
         self.name = 0;
     }
 }
@@ -159,11 +147,7 @@ impl Missing {
 
     /// Add a missing component index to the set.
     #[inline]
-    #[warn(
-        clippy::missing_const_for_fn,
-        reason = "can be made const with MSRV 1.83"
-    )]
-    pub fn insert(&mut self, ix: usize) {
+    pub const fn insert(&mut self, ix: usize) {
         self.0 |= Self::single(ix).0;
     }