+++ CHANGELOG.md
@@ -13,7 +13,11 @@ You can find its changes [documented below](#033-2026-05-05).
 
 ## [Unreleased][]
 
-This release has an [MSRV][] of 1.82.
+This release has an [MSRV][] of 1.86.
+
+### Changed
+
+- `Flags::set_missing`, `Flags::discard_name`, `Flags::set_named_color_space`, and `Missing::insert` are now `const` thanks to the MSRV update. ([#xxx] by [@DJMcNab][])
 
 ## [0.3.3][] (2026-05-05)